Run createProof as a job in a child process
